requirements.txt
Django==1.11
django-tinymce4-lite==1.7.4
django-forms-bootstrap==3.1.0
django-crispy-forms==1.8.0
django-allauth==0.40.0
dj-database-url==0.5.0
django-polymorphic==2.1.2
django-taggit==0.24.0
pytz==2019.2
chardet==3.0.4
colorama==0.3.7
docutils==0.14
jmespath==0.9.3
olefile==0.45.1
Pillow==5.1.0
python-dateutil==2.6.1
psycopg2-binary==2.8.3
requests==2.18.4
six==1.11.0
urllib3==1.22